requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests
azure-functions
azure-storage-blob
PyPDF2
reportlab
docx2pdf
python-docx
psycopg2-binary